refactor() split functions.py files into one file per helper
