+++ flashback/importing/__init__.py
@@ -1,5 +1,5 @@
-from .functions import import_class_from_path
-from .functions import import_module_from_path
+from .import_class_from_path import import_class_from_path
+from .import_module_from_path import import_module_from_path
 
 
 __all__ = (
